fix: test class rename
