+++ arc-ws/src/test/java/fr/insee/arc/ws/services/importServlet/ImportStep1OkTest.java
@@ -16,13 +16,9 @@ import fr.insee.arc.utils.dao.UtilitaireDao;
 import fr.insee.arc.utils.exception.ArcException;
 import fr.insee.arc.utils.query.InitializeQueryTest;
 import fr.insee.arc.ws.services.importServlet.actions.SendResponse;
-import fr.insee.arc.ws.services.importServlet.bo.ExportTrackingType;
 
-public class ImportStep1InitializeClientTablesServiceTestKO extends ServletArc {
-	
-	/**
-	 * 
-	 */
+public class ImportStep1OkTest extends ServletArc {
+
 	private static final long serialVersionUID = -7832574224892526397L;
 
 
@@ -51,21 +47,31 @@ public class ImportStep1InitializeClientTablesServiceTestKO extends ServletArc {
 	}
 	
 	
+	@Test(expected = ArcException.class)
+	public void testExecuteFamilyNotValid() throws ArcException {
+		JSONObject clientJsonInput = new JSONObject(
+				"{\"familleNorme\":\"RESIL\",\"periodicite\":\"M\",\"service\":\"arcClient\",\"validiteSup\":\"2032-03-01\",\"format\":\"csv_gzip\",\"reprise\":false,\"client\":\"ARTEMIS\",\"environnement\":\"arc_bas1\"}");
+		executeImportStep1(clientJsonInput);
+	}
+	
+	
 	@Test
 	public void testExecute() throws ArcException {
 
 		JSONObject clientJsonInput = new JSONObject(
 				"{\"familleNorme\":\"DSN\",\"periodicite\":\"M\",\"service\":\"arcClient\",\"validiteSup\":\"2032-03-01\",\"format\":\"csv_gzip\",\"reprise\":false,\"client\":\"ARTEMIS\",\"environnement\":\"arc_bas1\"}");
 
-		String arcResponse = executeImportStep1(clientJsonInput);
+		executeImportStep1(clientJsonInput);
 
-		testCreateAndDropWsPending(arcResponse);
-		
-		testCreateTableWsKo(arcResponse);
-		
+		testCreateAndDropWsPending();
+		testCreateTableNmcl();
+		testCreateTableVarMetier();
+		testCreateTableTableMetier();
+		testCreateTableTableFamille();
+		testCreateTableTablePeriodicite();
 	}
 	
-	private void testCreateAndDropWsPending(String arcResponse) throws ArcException {
+	private void testCreateAndDropWsPending() throws ArcException {
 		
 		// check that the parallel thread that create tables drop the table ws_pending
 
@@ -73,7 +79,7 @@ public class ImportStep1InitializeClientTablesServiceTestKO extends ServletArc {
 		int maxIteration = 50;
 		int i=0;
 		
-		while (i<maxIteration && UtilitaireDao.get(0).isTableExiste(InitializeQueryTest.c, arcResponse+"_ws_pending"))
+		while (i<maxIteration && UtilitaireDao.get(0).isTableExiste(InitializeQueryTest.c, "arc_bas1.ARTEMIS_%_ws_pending"))
 		{
 			i++;
 			UtilitaireDao.get(0).executeImmediate(InitializeQueryTest.c, "SELECT pg_sleep(1);");
@@ -83,13 +89,32 @@ public class ImportStep1InitializeClientTablesServiceTestKO extends ServletArc {
 		assertTrue(i<maxIteration);
 	}
 	
-	private void testCreateTableWsKo(String arcResponse) throws ArcException {
-		ArcPreparedStatementBuilder query = new ArcPreparedStatementBuilder();
-		
-		query.append("SELECT 1 FROM "+arcResponse+"_ws_tracking where tracking_type="+query.quoteText(ExportTrackingType.KO.toString()));
-		
-		UtilitaireDao.get(0).hasResults(InitializeQueryTest.c, query);
+	private void testCreateTableNmcl() throws ArcException {
+		// table image created should be like arc_bas1.ARTEMIS_timestamp_<tablename_to_retrieve>
+		assertTrue(UtilitaireDao.get(0).isTableExiste(InitializeQueryTest.c, "arc_bas1.ARTEMIS_%_nmcl_table1"));
+		assertTrue(UtilitaireDao.get(0).isTableExiste(InitializeQueryTest.c, "arc_bas1.ARTEMIS_%_nmcl_table2"));
 	}
+
+	private void testCreateTableVarMetier() throws ArcException {
+		// table image created should be like arc_bas1.ARTEMIS_timestamp_<tablename_to_retrieve>
+		assertTrue(UtilitaireDao.get(0).isTableExiste(InitializeQueryTest.c, "arc_bas1.ARTEMIS_%_mod_variable_metier"));
+	}
+	
+	private void testCreateTableTableMetier() throws ArcException {
+		// table image created should be like arc_bas1.ARTEMIS_timestamp_<tablename_to_retrieve>
+		assertTrue(UtilitaireDao.get(0).isTableExiste(InitializeQueryTest.c, "arc_bas1.ARTEMIS_%_mod_table_metier"));
+	}
+	
+	private void testCreateTableTableFamille() throws ArcException {
+		// table image created should be like arc_bas1.ARTEMIS_timestamp_<tablename_to_retrieve>
+		assertTrue(UtilitaireDao.get(0).isTableExiste(InitializeQueryTest.c, "arc_bas1.ARTEMIS_%_ext_mod_famille"));
+	}
+	
+	private void testCreateTableTablePeriodicite() throws ArcException {
+		// table image created should be like arc_bas1.ARTEMIS_timestamp_<tablename_to_retrieve>
+		assertTrue(UtilitaireDao.get(0).isTableExiste(InitializeQueryTest.c, "arc_bas1.ARTEMIS_%_ext_mod_periodicite"));
+	}
+	
 	
 	/**
 	 * initialize data for the tests
@@ -99,7 +124,6 @@ public class ImportStep1InitializeClientTablesServiceTestKO extends ServletArc {
 	private static void initializeTestData() throws SQLException, ArcException {
 		
 		ArcPreparedStatementBuilder query;
-
 		query = new ArcPreparedStatementBuilder();
 
 		query.append("CREATE SCHEMA arc;");
@@ -108,7 +132,8 @@ public class ImportStep1InitializeClientTablesServiceTestKO extends ServletArc {
 		
 		// family and client tables
 		query.append("CREATE TABLE arc.ihm_client AS ");
-		query.append("SELECT 'DSN' as id_famille,'ARTEMIS' as id_application UNION ALL ");
+		query.append("SELECT 'DSN' as id_famille,'ARTEMIS' as id_application");
+		query.append(SQL.UNION_ALL);
 		query.append("SELECT 'DSN' as id_famille,'DSNFLASH' as id_application");
 		query.append(SQL.END_QUERY);
 		
@@ -116,18 +141,20 @@ public class ImportStep1InitializeClientTablesServiceTestKO extends ServletArc {
 		query.append(SQL.END_QUERY);
 
 		query.append("CREATE TABLE arc_bas1.mod_table_metier AS ");
-		query.append("SELECT 'DSN' as id_famille,'mapping_dsn_test1_ok' as nom_table_metier UNION ALL ");
+		query.append("SELECT 'DSN' as id_famille,'mapping_dsn_test1_ok' as nom_table_metier");
+		query.append(SQL.UNION_ALL);
 		query.append("SELECT 'PASRAU' as id_famille,'mapping_pasrau_test_ok' as nom_table_metier");
 		query.append(SQL.END_QUERY);
 		
-		// table variable_metier doesn't exists, it will crash
+		query.append("CREATE TABLE arc_bas1.mod_variable_metier AS SELECT 'DSN' as id_famille, 'mapping_dsn_test1_ok' as nom_table_metier, 'id_source' as nom_variable_metier");
+		query.append(SQL.END_QUERY);
 
 		// pilotage tables
 		query.append("CREATE TABLE arc_bas1.pilotage_fichier AS ");
 		query.append("SELECT 'file_to_retrieve.xml' as id_source, 'PHASE3V1' as id_norme, '2023-10-01' as validite,'M' as periodicite");
 		query.append(", 'MAPPING' as phase_traitement, '{OK}'::text[] as etat_traitement, '2023-11-30 10:29:47.000'::timestamp as date_traitement");
 		query.append(", null::text[] as client, null::timestamp[] as date_client");
-		query.append(" UNION ALL ");
+		query.append(SQL.UNION_ALL);
 		// file that mustn't be retrieved when reprise is false and family is DSN
 		query.append("SELECT 'file_not_to_retrieve_when_reprise_false.xml' as id_source, 'PHASE3V1' as id_norme, '2023-10-01' as validite,'M' as periodicite");
 		query.append(", 'MAPPING' as phase_traitement, '{OK}'::text[] as etat_traitement, '2023-11-30 10:29:47.000'::timestamp as date_traitement");
